Move tile config to its own file as well
